Int64 to Int

+++ src/params_intervals.jl
@@ -17,7 +17,7 @@ end
 "Structure storing input data for parameter interval calculation"
 struct ParamInput
     init_params::Vector{Float64} # initial parameters vector
-    id::Int64 # id of the parameter for analysis
+    id::Int # id of the parameter for analysis
     loss_crit::Float64 # loss function maximum value, "identifiability level"
     loss_func::Function # loss function
     method::Symbol # method ONE_PASS, D2D_PLE
@@ -25,7 +25,7 @@ struct ParamInput
     scan_bound::Vector{Float64} # search bounds for id parameter (default - [1e-9,1e9])
     bounds::Vector{Vector{Float64}} # bound constraints for all parameters except id
     local_alg::Symbol # local fitting algorithm (default - :LN_NELDERMEAD)
-    max_iter::Int64 # maximum function evaluations
+    max_iter::Int # maximum function evaluations
     ptol::Float64 # fitting tolerance for local optimizer (default - 1e-3)
     losstol::Float64 # constraints tolerance
 end
@@ -35,14 +35,14 @@ struct ParamInterval
     input::ParamInput # input data
     interval::Array{Float64, 1} # result of interval calculation. If open intervals than undefined
     ret_codes::Array{Symbol, 1} # returned result: :BOUNDS_REACHED if cannot calculate, :FTOL_REACHED if everything ok
-    count_evals::Array{Int64, 1} # count of loss_func calls
+    count_evals::Array{Int, 1} # count of loss_func calls
     loss_final::Array{Float64, 1} # value of loss_func calculated on intervals or scan_bound
 
     profile_buffer::Array{ProfilePoint, 1} # storage for true points of loss_func profile
 end
 
 """
-    params_intervals(init_params::Vector{Float64}, id::Int64,
+    params_intervals(init_params::Vector{Float64}, id::Int,
     loss_crit::Float64, loss_func::Function; <keyword arguments>)
 
 Computes confidence interval for `id` parameter of `init_params` vector and `loss_func`
@@ -57,13 +57,13 @@ Returns `ParamInterval` structure storing all input data and estimated confidenc
 - `scan_bound::Vector{Float64}`: search bounds for id parameter (default `[-9.,9.]`).
 - `local_alg::Symbol`: fitting algorithm (default `:LN_NELDERMEAD`).
 - `bounds::Vector{Vector{Float64}}`: bound constraints for all parameters (default `[-Inf,Inf]`).
-- `max_iter::Int64`: maximum `loss_func` evaluations (default `10^5`).
+- `max_iter::Int`: maximum `loss_func` evaluations (default `10^5`).
 - `ptol::Float64`: fitting tolerance for optimizer (default `1e-3`).
 - `losstol::Float64`: constraints tolerance (default `1e-3`).
 """
 function params_intervals(
     init_params::Vector{Float64},
-    id::Int64,
+    id::Int,
     loss_crit::Float64,
     loss_func::Function; # (Array{Float64,1})
 
@@ -80,7 +80,7 @@ function params_intervals(
         logscale
     ),
     local_alg::Symbol = :LN_NELDERMEAD,
-    max_iter::Int64 = 100000,
+    max_iter::Int = 100000,
     ptol::Float64 = 1e-3,
     losstol::Float64 = 1e-3, # tol_const::Float64 = 1e-3
     # ftol_glob::Float64 = 0.
